+++ apps/picsa-tools/manual-tool/src/app/pages/home/home.module.ts
@@ -1,9 +1,9 @@
+/* eslint-disable @nrwl/nx/enforce-module-boundaries */
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { Route, RouterModule } from '@angular/router';
-import { PicsaTranslateModule } from '@picsa/shared/modules';
-
 import { ResourcesComponentsModule } from '@picsa/resources/src/app/components/components.module';
+import { PicsaTranslateModule } from '@picsa/shared/modules';
 
 import { ManualToolComponentsModule } from '../../components/components.module';
 import { HomeComponent } from './home.component';
